+++ gatsby-node.js
@@ -169,16 +169,6 @@ exports.onCreateWebpackConfig = ({ stage, loaders, actions }) => {
           },
         ],
       },
-      resolve: {
-        fallback: {
-          assert: require.resolve("assert/"),
-          crypto: require.resolve("crypto-browserify"),
-          http: require.resolve("stream-http"),
-          https: require.resolve("https-browserify"),
-          os: require.resolve("os-browserify/browser"),
-          stream: require.resolve("stream-browserify"),
-        },
-      },
     })
   }
 }
