+++ package.json
@@ -40,7 +40,6 @@
     "react-compare-slider": "^2.2.0",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.48.2",
     "react-icons": "^4.3.1",
     "react-instantsearch-dom": "^6.23.4",
     "react-is": "^17.0.1",
@@ -62,11 +61,7 @@
     "@types/swiper": "6.0.0",
     "@typescript-eslint/eslint-plugin": "5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
-    "assert": "^2.1.0",
     "babel-eslint": "^10.1.0",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "crypto-browserify": "^3.12.0",
     "eslint": "^8.10.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
@@ -78,15 +73,9 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.3",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "https-browserify": "^1.0.0",
-    "os-browserify": "^0.3.0",
     "prettier": "^2.5.1",
-    "process": "^0.11.10",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
     "ts-node": "10.7.0",
     "typescript": "4.6.2",
-    "url": "^0.11.3",
     "webpack": "^5.70.0"
   },
   "keywords": [
