+++ src/contexts/reviews/ReviewsProvider.tsx
@@ -13,11 +13,10 @@ import {
 
 type Props = {
   productId: number
-  productHandle: string
   children: ReactNode | ReactNode[]
 }
 
-export function ReviewsProvider({ productId, productHandle, children }: Props) {
+export function ReviewsProvider({ productId, children }: Props) {
   const [data, setData] = useState<YotpoRetrieveReviewsResponse | null>(null)
   const [isLoading, setIsLoading] = useState(true)
   const [isRefetching, setIsRefetching] = useState(false)
@@ -109,7 +108,6 @@ export function ReviewsProvider({ productId, productHandle, children }: Props) {
   const reviewsContextValue = useMemo(
     () => ({
       data,
-      productHandle,
       isLoading,
       mutateReviewThumbVote,
       refreshToPage,
