+++ src/contexts/reviews/types.d.ts
@@ -4,7 +4,6 @@ export type ReviewContextType = {
   isLoading: boolean
   isRefetching: boolean
   data: YotpoRetrieveReviewsResponse | null
-  productHandle: string
   mutateReviewThumbVote: (props: {
     vote: "up" | "down"
     reviewId: number
